# api/openapi/q2-api.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "q2 API",
    "description": "Kudos (q2) — shared self-care goals. Reference API for the initial version.",
    "version": "v1"
  },
  "paths": {
    "/api/profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Returns who you are, your streak, your badges and your recent activity.",
        "operationId": "GetProfile",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Profile"
        ],
        "summary": "Changes your display name or your picture. Omitted properties keep their value.",
        "operationId": "UpdateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/people/{personId}": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Returns somebody else's profile, with a balance covering only the goals you share.",
        "operationId": "GetPersonProfile",
        "parameters": [
          {
            "name": "personId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonProfileResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/goals/{id}/proof": {
      "post": {
        "tags": [
          "Proofs"
        ],
        "summary": "Delivers a photograph into the goal's open window for its friends to vote on.",
        "operationId": "SubmitGoalProof",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitProofRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProofResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/proofs/{id}": {
      "get": {
        "tags": [
          "Proofs"
        ],
        "summary": "Returns one photograph, if it is yours to see.",
        "operationId": "GetProof",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProofResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/proofs/{id}/vote": {
      "post": {
        "tags": [
          "Proofs"
        ],
        "summary": "Confirms or doubts a photograph. One say per person, and it stands.",
        "operationId": "VoteOnProof",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CastVoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProofResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/proofs/{id}/reactions": {
      "post": {
        "tags": [
          "Proofs"
        ],
        "summary": "Adds or takes back a reaction. The same kind twice takes it back.",
        "operationId": "ReactToProof",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactToProofRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProofResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivityKind": {
        "enum": [
          "TaskCompleted",
          "StreakReached",
          "GoalProgress",
          "GoalCreated",
          "WindowAtRisk"
        ]
      },
      "ActivityResponse": {
        "required": [
          "id",
          "actor",
          "kind",
          "subject",
          "amount",
          "kudosCount",
          "hasMyKudos",
          "occurredAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "actor": {
            "$ref": "#/components/schemas/PersonSummary"
          },
          "kind": {
            "$ref": "#/components/schemas/ActivityKind"
          },
          "subject": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kudosCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasMyKudos": {
            "type": "boolean"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BadgeKey": {
        "enum": [
          "StreakHero",
          "EarlyBird",
          "Bookworm",
          "KudosGiver",
          "Marathon",
          "WeeklyWinner"
        ]
      },
      "BadgeResponse": {
        "required": [
          "key",
          "isEarned",
          "earnedOn"
        ],
        "type": "object",
        "properties": {
          "key": {
            "$ref": "#/components/schemas/BadgeKey"
          },
          "isEarned": {
            "type": "boolean"
          },
          "earnedOn": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        }
      },
      "BalanceResponse": {
        "required": [
          "done",
          "missed"
        ],
        "type": "object",
        "properties": {
          "done": {
            "type": "integer",
            "format": "int32"
          },
          "missed": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CastVoteRequest": {
        "type": "object",
        "properties": {
          "value": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/VoteValue"
              }
            ]
          }
        }
      },
...
      "DaySummaryResponse": {
        "required": [
          "done",
          "total",
          "percent"
        ],
        "type": "object",
        "properties": {
          "done": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "percent": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "FriendshipState": {
        "enum": [
          "None",
          "RequestSent",
          "RequestReceived",
          "Friends",
          "Self"
        ]
      },
...
      "GoalInstanceResponse": {
        "required": [
          "id",
          "startsOn",
          "dueOn",
          "dueAt",
          "requiredProofs",
          "confirmedProofs",
          "remainingProofs",
          "status",
          "pendingProofId",
          "acceptsProof"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "startsOn": {
            "type": "string",
            "format": "date"
          },
          "dueOn": {
            "type": "string",
            "format": "date"
          },
          "dueAt": {
            "type": "string",
            "format": "date-time"
          },
          "requiredProofs": {
            "type": "integer",
            "format": "int32"
          },
          "confirmedProofs": {
            "type": "integer",
            "format": "int32"
          },
          "remainingProofs": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/GoalInstanceStatus"
          },
          "pendingProofId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "acceptsProof": {
            "type": "boolean"
          }
        }
      },
...
      "GoalPauseResponse": {
        "required": [
          "id",
          "reason",
          "startsOn",
          "endsOn",
          "endsAt",
          "days",
          "vetoCount",
          "vetoesRequired",
          "vetoedByMe",
          "canVeto"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string"
          },
          "startsOn": {
            "type": "string",
            "format": "date"
          },
          "endsOn": {
            "type": "string",
            "format": "date"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "days": {
            "type": "integer",
            "format": "int32"
          },
          "vetoCount": {
            "type": "integer",
            "format": "int32"
          },
          "vetoesRequired": {
            "type": "integer",
            "format": "int32"
          },
          "vetoedByMe": {
            "type": "boolean"
          },
          "canVeto": {
            "type": "boolean"
          }
        }
      },
      "GoalResponse": {
        "required": [
          "id",
          "title",
          "description",
          "icon",
          "schedule",
          "status",
          "isGroup",
          "current",
          "streak",
          "windowsDone",
          "windowsMissed",
          "reminderAt",
          "targetDate",
          "createdAt",
          "closedAt",
          "participants",
          "isOverdue",
          "isMine",
          "risk",
          "pause",
          "remainingPauses"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string"
          },
          "schedule": {
            "$ref": "#/components/schemas/GoalScheduleResponse"
          },
          "status": {
            "$ref": "#/components/schemas/GoalStatus"
          },
          "isGroup": {
            "type": "boolean"
          },
          "current": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/GoalInstanceResponse"
              }
            ]
          },
          "streak": {
            "type": "integer",
            "format": "int32"
          },
          "windowsDone": {
            "type": "integer",
            "format": "int32"
          },
          "windowsMissed": {
            "type": "integer",
            "format": "int32"
          },
          "reminderAt": {
            "type": "string",
            "format": "time",
            "nullable": true
          },
          "targetDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonSummary"
            }
          },
          "isOverdue": {
            "type": "boolean"
          },
          "isMine": {
            "type": "boolean"
          },
          "risk": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/RiskResponse"
              }
            ]
          },
          "pause": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/GoalPauseResponse"
              }
            ]
          },
          "remainingPauses": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
...
      "GoalScheduleResponse": {
        "required": [
          "kind",
          "everyDays",
          "weekdays",
          "times",
          "period"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ScheduleKind"
          },
          "everyDays": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "weekdays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Weekday"
            }
          },
          "times": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "period": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/QuotaPeriod"
              }
            ]
          }
        }
      },
      "GoalStatus": {
        "enum": [
          "Active",
          "Completed",
          "Archived"
        ]
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "KudosKind": {
        "enum": [
          "Fire",
          "Strong",
          "Applause"
        ]
      },
...
      "PersonProfileResponse": {
        "required": [
          "person",
          "state",
          "streak",
          "kudosReceived",
          "goalsCompleted",
          "balance",
          "sharedGoals"
        ],
        "type": "object",
        "properties": {
          "person": {
            "$ref": "#/components/schemas/PersonSummary"
          },
          "state": {
            "$ref": "#/components/schemas/FriendshipState"
          },
          "streak": {
            "type": "integer",
            "format": "int32"
          },
          "kudosReceived": {
            "type": "integer",
            "format": "int32"
          },
          "goalsCompleted": {
            "type": "integer",
            "format": "int32"
          },
          "balance": {
            "$ref": "#/components/schemas/BalanceResponse"
          },
          "sharedGoals": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PersonSummary": {
        "required": [
          "id",
          "displayName",
          "handle",
          "initials",
          "avatarColor",
          "isOnline",
          "avatarImageId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "handle": {
            "type": "string"
          },
          "initials": {
            "type": "string"
          },
          "avatarColor": {
            "type": "string"
          },
          "isOnline": {
            "type": "boolean"
          },
          "avatarImageId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ProfileResponse": {
        "required": [
          "person",
          "streak",
          "kudosReceived",
          "goalsCompleted",
          "weekActivity",
          "today",
          "unreadChats",
          "pendingFriendRequests",
          "badges",
          "recentActivity",
          "balance",
          "atRisk"
        ],
        "type": "object",
        "properties": {
          "person": {
            "$ref": "#/components/schemas/PersonSummary"
          },
          "streak": {
            "type": "integer",
            "format": "int32"
          },
          "kudosReceived": {
            "type": "integer",
            "format": "int32"
          },
          "goalsCompleted": {
            "type": "integer",
            "format": "int32"
          },
          "weekActivity": {
            "type": "array",
            "items": {
              "type": "boolean"
            }
          },
          "today": {
            "$ref": "#/components/schemas/DaySummaryResponse"
          },
          "unreadChats": {
            "type": "integer",
            "format": "int32"
          },
          "pendingFriendRequests": {
            "type": "integer",
            "format": "int32"
          },
          "badges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BadgeResponse"
            }
          },
          "recentActivity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityResponse"
            }
          },
          "balance": {
            "$ref": "#/components/schemas/BalanceResponse"
          },
          "atRisk": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoalResponse"
            }
          }
        }
      },
      "ProofResponse": {
        "required": [
          "id",
          "goalId",
          "goalInstanceId",
          "uploader",
          "imageId",
          "status",
          "attempt",
          "attemptsLeft",
          "capturedInApp",
          "createdAt",
          "expiresAt",
          "votes",
          "reactions"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "goalId": {
            "type": "string",
            "format": "uuid"
          },
          "goalInstanceId": {
            "type": "string",
            "format": "uuid"
          },
          "uploader": {
            "$ref": "#/components/schemas/PersonSummary"
          },
          "imageId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/ProofStatus"
          },
          "attempt": {
            "type": "integer",
            "format": "int32"
          },
          "attemptsLeft": {
            "type": "integer",
            "format": "int32"
          },
          "capturedInApp": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "votes": {
            "$ref": "#/components/schemas/VoteSummaryResponse"
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReactionSummaryResponse"
            }
          }
        }
      },
      "ProofStatus": {
        "enum": [
          "Voting",
          "Confirmed",
          "Rejected"
        ]
      },
...
      "QuotaPeriod": {
        "enum": [
          "Week",
          "Month",
          null
        ]
      },
...
      "ReactToProofRequest": {
        "type": "object",
        "properties": {
          "kind": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/KudosKind"
              }
            ]
          }
        }
      },
      "ReactionSummaryResponse": {
        "required": [
          "kind",
          "count",
          "isMine"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/KudosKind"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "isMine": {
            "type": "boolean"
          }
        }
      },
...
      "RiskResponse": {
        "required": [
          "reason",
          "missingProofs",
          "requiredProofs",
          "remainingDays"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "$ref": "#/components/schemas/RiskReason"
          },
          "missingProofs": {
            "type": "integer",
            "format": "int32"
          },
          "requiredProofs": {
            "type": "integer",
            "format": "int32"
          },
          "remainingDays": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ScheduleKind": {
        "enum": [
          "Once",
          "Interval",
          "Weekdays",
          "Times"
        ]
      },
...
      "SubmitProofRequest": {
        "type": "object",
        "properties": {
          "imageId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "capturedInApp": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarImageId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
...
      "VoteSummaryResponse": {
        "required": [
          "confirmCount",
          "doubtCount",
          "confirmedBy",
          "myVote",
          "canIVote"
        ],
        "type": "object",
        "properties": {
          "confirmCount": {
            "type": "integer",
            "format": "int32"
          },
          "doubtCount": {
            "type": "integer",
            "format": "int32"
          },
          "confirmedBy": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonSummary"
            }
          },
          "myVote": {
            "oneOf": [
              {
                "enum": [
                  null
                ]
              },
              {
                "$ref": "#/components/schemas/VoteValue"
              }
            ]
          },
          "canIVote": {
            "type": "boolean"
          }
        }
      },
      "VoteValue": {
        "enum": [
          "Confirm",
          "Doubt",
          null
        ]
      },
...
    }
  },
...
}